requirements.txt
kafka-python-ng==2.2.2
requests==2.31.0
nltk==3.8.1
flask==3.0.0
python-dotenv==1.0.0
